+++ certlogic/certlogic-dart/lib/src/evaluate.dart
@@ -1,12 +1,20 @@
 import 'package:certlogic_dart/src/internals.dart';
 import 'package:certlogic_dart/src/typings.dart';
 
+class CertLogicException implements Exception {
+  final String message;
+
+  const CertLogicException(this.message);
+
+  String toString() => 'CertLogicException: $message';
+}
+
 /// The library, you can use the static functions, so you don't need to instantiate this class
 class CertLogic {
   /// Evaluate a variable
   static dynamic _evaluateVar(dynamic value, dynamic data) {
     if (value is! String) {
-      throw Exception("not of the form { 'var': '<path>' }");
+      throw CertLogicException("not of the form { 'var': '<path>' }");
     }
     if (value == '') return data;
     var returnData = data;
@@ -33,7 +41,8 @@ class CertLogic {
     return returnData;
   }
 
-  static dynamic _evaluateIf(dynamic guard, dynamic then, dynamic elseDo, dynamic data) {
+  static dynamic _evaluateIf(
+      dynamic guard, dynamic then, dynamic elseDo, dynamic data) {
     final dynamic evalGuard = evaluate(guard, data);
     if (CertLogicInternals.isTruthy(evalGuard)) {
       return evaluate(then, data);
@@ -42,7 +51,18 @@ class CertLogic {
     }
   }
 
-  static bool Function(dynamic l, dynamic r) _compareFunctionFor(String operatorAsString) {
+  static String? _evaluateExtractFromUVCI(
+      dynamic operand, int index, dynamic data) {
+    final evalOperand = evaluate(operand, data);
+    if (evalOperand is! String) {
+      throw CertLogicException(
+          '"UVCI" argument (#1) of "extractFromUVCI" must be either a string or null');
+    }
+    return CertLogicInternals.extractFromUVCI(evalOperand, index);
+  }
+
+  static bool Function(dynamic l, dynamic r) _compareFunctionFor(
+      String operatorAsString) {
     return (dynamic l, dynamic r) {
       num left = 0;
       if (l is DateTime) {
@@ -50,7 +70,7 @@ class CertLogic {
       } else if (l is num) {
         left = l;
       } else {
-        throw Exception('incorrect type');
+        throw CertLogicException('incorrect type');
       }
       num right = 0;
       if (r is DateTime) {
@@ -58,7 +78,7 @@ class CertLogic {
       } else if (r is num) {
         right = r;
       } else {
-        throw Exception('incorrect type');
+        throw CertLogicException('incorrect type');
       }
       switch (operatorAsString) {
         case '<':
@@ -70,7 +90,7 @@ class CertLogic {
         case '>=':
           return left >= right;
         default:
-          throw Exception('incorrect operator');
+          throw CertLogicException('incorrect operator');
       }
     };
   }
@@ -81,13 +101,16 @@ class CertLogic {
       case 2:
         return compFunc(values.elementAt(0), values.elementAt(1));
       case 3:
-        return compFunc(values.elementAt(0), values.elementAt(1)) && compFunc(values.elementAt(1), values.elementAt(2)!);
+        return compFunc(values.elementAt(0), values.elementAt(1)) &&
+            compFunc(values.elementAt(1), values.elementAt(2)!);
       default:
-        throw Exception("invalid number of operands to a '$operatorAsString' operation");
+        throw CertLogicException(
+            "invalid number of operands to a '$operatorAsString' operation");
     }
   }
 
-  static String _comparisonOperatorForDateTimeComparison(String operatorAsString) {
+  static String _comparisonOperatorForDateTimeComparison(
+      String operatorAsString) {
     switch (operatorAsString) {
       case 'after':
         return '>';
@@ -98,15 +121,17 @@ class CertLogic {
       case 'not-before':
         return '>=';
       default:
-        throw Exception('invalid operator');
+        throw CertLogicException('invalid operator');
     }
   }
 
-  static dynamic _evaluateInfix(String operatorAsString, Iterable values, dynamic data) {
+  static dynamic _evaluateInfix(
+      String operatorAsString, Iterable values, dynamic data) {
     switch (operatorAsString) {
       case 'and':
         if (values.length < 2) {
-          throw Exception("an 'and' operation must have at least 2 operands");
+          throw CertLogicException(
+              "an 'and' operation must have at least 2 operands");
         }
         break;
       case '<':
@@ -117,10 +142,14 @@ class CertLogic {
       case 'before':
       case 'not-after':
       case 'not-before':
-        if (values.length < 2 || values.length > 3) throw Exception("an operation with operator '$operatorAsString' must have 2 or 3 operands");
+        if (values.length < 2 || values.length > 3)
+          throw CertLogicException(
+              "an operation with operator '$operatorAsString' must have 2 or 3 operands");
         break;
       default:
-        if (values.length != 2) throw Exception("an operation with operator '$operatorAsString' must have 2 operands");
+        if (values.length != 2)
+          throw CertLogicException(
+              "an operation with operator '$operatorAsString' must have 2 operands");
         break;
     }
     final evalArgs = values.map<dynamic>((dynamic arg) => evaluate(arg, data));
@@ -132,7 +161,8 @@ class CertLogic {
           dynamic r = evalArgs.elementAt(1);
           if (r is Map) r = r.keys;
           if (r is! Iterable) {
-            throw Exception("right-hand side of an 'in' operation must be an array");
+            throw CertLogicException(
+                "right-hand side of an 'in' operation must be an array");
           }
           return r.contains(evalArgs.elementAt(0));
         }
@@ -141,7 +171,8 @@ class CertLogic {
           final dynamic l = evalArgs.elementAt(0);
           final dynamic r = evalArgs.elementAt(1);
           if (l is! num || r is! num) {
-            throw Exception('operands of this operation must both be integers');
+            throw CertLogicException(
+                'operands of this operation must both be integers');
           }
           return l + r;
         }
@@ -154,7 +185,8 @@ class CertLogic {
       case '>=':
         {
           if (!evalArgs.every((dynamic e) => e is num)) {
-            throw Exception('all operands of a comparison operation must be of integer type');
+            throw CertLogicException(
+                'all operands of a comparison operation must be of integer type');
           }
           return _compare(operatorAsString, evalArgs);
         }
@@ -164,12 +196,16 @@ class CertLogic {
       case 'not-before':
         {
           if (!evalArgs.every((dynamic e) => e is DateTime)) {
-            throw Exception('all operands of a date-time comparison must be date-times');
+            throw CertLogicException(
+                'all operands of a date-time comparison must be date-times');
           }
-          return _compare(_comparisonOperatorForDateTimeComparison(operatorAsString), evalArgs);
+          return _compare(
+              _comparisonOperatorForDateTimeComparison(operatorAsString),
+              evalArgs);
         }
       default:
-        throw Exception("unhandled infix operator '$operatorAsString'");
+        throw CertLogicException(
+            "unhandled infix operator '$operatorAsString'");
     }
   }
 
@@ -181,36 +217,44 @@ class CertLogic {
     if (CertLogicInternals.isTruthy(operand)) {
       return false;
     }
-    throw Exception('operand of ! _evaluates to something neither truthy, nor falsy: $operand');
+    throw CertLogicException(
+        'operand of ! _evaluates to something neither truthy, nor falsy: $operand');
   }
 
-  static DateTime _evaluatePlusTime(dynamic dateOperand, dynamic amount, String unit, dynamic data) {
+  static DateTime _evaluatePlusTime(
+      dynamic dateOperand, dynamic amount, String unit, dynamic data) {
     if (amount is! num) {
-      throw Exception("'amount' argument (#2) of 'plusTime' must be an integer");
+      throw CertLogicException(
+          "'amount' argument (#2) of 'plusTime' must be an integer");
     }
     if (!['day', 'hour', 'month', 'year'].contains(unit)) {
-      throw Exception("'unit' argument (#3) of 'plusTime' must be a string 'year', 'month', 'day' or 'hour'");
+      throw CertLogicException(
+          "'unit' argument (#3) of 'plusTime' must be a string 'year', 'month', 'day' or 'hour'");
     }
     final dynamic dateTimeStr = evaluate(dateOperand, data);
     if (dateTimeStr is! String) {
-      throw Exception("date argument of 'plusTime' must be a string");
+      throw CertLogicException("date argument of 'plusTime' must be a string");
     }
-    return CertLogicInternals.plusTime(dateTimeStr, amount.toInt(), CertLogicTimeUnits.fromString(unit));
+    return CertLogicInternals.plusTime(
+        dateTimeStr, amount.toInt(), CertLogicTimeUnits.fromString(unit));
   }
 
-  static dynamic _evaluateReduce(dynamic operand, dynamic lambda, dynamic initial, dynamic data) {
+  static dynamic _evaluateReduce(
+      dynamic operand, dynamic lambda, dynamic initial, dynamic data) {
     final dynamic evalOperand = evaluate(operand, data);
     dynamic evalInitial() => evaluate(initial, data);
     if (evalOperand == null) {
       return evalInitial();
     }
     if (evalOperand is! Iterable) {
-      throw Exception('operand of reduce _evaluated to a non-null non-array');
+      throw CertLogicException(
+          'operand of reduce _evaluated to a non-null non-array');
     }
 
     var accumulator = evalInitial();
     evalOperand.forEach((dynamic current) {
-      accumulator = evaluate(lambda, {'current': current, 'accumulator': accumulator});
+      accumulator =
+          evaluate(lambda, {'current': current, 'accumulator': accumulator});
     });
     return accumulator;
   }
@@ -222,7 +266,7 @@ class CertLogic {
       return expr;
     }
     if (expr == null) {
-      throw Exception('invalid CertLogic expression: $expr');
+      throw CertLogicException('invalid CertLogic expression: $expr');
     }
     if (expr is Iterable) {
       return expr.map<dynamic>((dynamic item) => evaluate(item, data));
@@ -230,7 +274,7 @@ class CertLogic {
     if (expr is Map) {
       final keys = expr.keys;
       if (keys.length != 1) {
-        throw Exception('unrecognised expression object encountered');
+        throw CertLogicException('unrecognised expression object encountered');
       }
       final operatorAsString = keys.first as String;
       final dynamic values = expr[operatorAsString];
@@ -238,7 +282,8 @@ class CertLogic {
         return _evaluateVar(values, data);
       }
       if (values is! List || values.isEmpty) {
-        throw Exception("operation not of the form { '<operator>': [ <values...> ] }");
+        throw CertLogicException(
+            "operation not of the form { '<operator>': [ <values...> ] }");
       }
       if (operatorAsString == 'if') {
         final dynamic guard = values[0];
@@ -246,20 +291,39 @@ class CertLogic {
         final dynamic elseDo = values[2];
         return _evaluateIf(guard, then, elseDo, data);
       }
-      if (['===', 'and', '>', '<', '>=', '<=', 'in', '+', 'after', 'before', 'not-after', 'not-before'].contains(operatorAsString)) {
+      if ([
+        '===',
+        'and',
+        '>',
+        '<',
+        '>=',
+        '<=',
+        'in',
+        '+',
+        'after',
+        'before',
+        'not-after',
+        'not-before'
+      ].contains(operatorAsString)) {
         return _evaluateInfix(operatorAsString, values, data);
       }
       if (operatorAsString == '!') {
         return _evaluateNot(values[0], data);
       }
       if (operatorAsString == 'plusTime') {
-        return _evaluatePlusTime(values[0], values[1], values[2] as String, data);
+        return _evaluatePlusTime(
+            values[0], values[1], values[2] as String, data);
       }
       if (operatorAsString == 'reduce') {
         return _evaluateReduce(values[0], values[1], values[2], data);
       }
-      throw Exception("unrecognised operator: '$operatorAsString'");
+      if (operatorAsString == 'extractFromUVCI' ||
+          operatorAsString == 'extractFromUCI') {
+        return _evaluateExtractFromUVCI(
+            values.elementAt(0), values.elementAt(1), data);
+      }
+      throw CertLogicException("unrecognised operator: '$operatorAsString'");
     }
-    throw Exception('invalid CertLogic expression: $expr');
+    throw CertLogicException('invalid CertLogic expression: $expr');
   }
 }