+++ certlogic/certlogic-dart/test/test_suite_test.dart
@@ -15,7 +15,9 @@ void main() {
 
   test('validate test suite', () async {
     final entries = dataDir.listSync(recursive: true).toList();
-    final files = entries.where((element) => element.path.endsWith('.json')).map((element) => File(element.path));
+    final files = entries
+        .where((element) => element.path.endsWith('.json'))
+        .map((element) => File(element.path));
     final failedNames = [];
     var success = 0;
     for (final file in files) {
@@ -27,11 +29,22 @@ void main() {
         testCase.assertions.forEach((assertion) {
           if (assertion.directive == TestDirective.SKIP) return;
           try {
-            final result = CertLogic.evaluate(assertion.certLogicExpression ?? testCase.certLogicExpression, assertion.data);
+            var result;
+            try {
+              result = CertLogic.evaluate(
+                  assertion.certLogicExpression ?? testCase.certLogicExpression,
+                  assertion.data);
+            } catch (e) {
+              if (e is CertLogicException) {
+                result = null;
+              } else {
+                rethrow;
+              }
+            }
             expect(result, assertion.expected);
             success++;
           } catch (e) {
-            failedNames.add(assertion.message ?? testCase.name);
+            failedNames.add('${testCase.name} - ${assertion.message}');
           }
         });
       });